Move CIO raw server to backend package
