+++ ktor-http/ktor-http-cio/jvm/test/io/ktor/tests/http/cio/TestHttpServer.kt
@@ -7,6 +7,7 @@ package io.ktor.tests.http.cio
 import io.ktor.http.cio.*
 import io.ktor.http.cio.internals.*
 import io.ktor.http.cio.internals.WeakTimeoutQueue
+import io.ktor.server.cio.backend.*
 import kotlinx.coroutines.*
 import io.ktor.utils.io.*
 import java.net.*
@@ -113,7 +114,11 @@ private suspend fun client(
     val timeouts = WeakTimeoutQueue(TimeUnit.HOURS.toMillis(1000))
 
     CoroutineScope(ioCoroutineContext + Dispatchers.Unconfined).startServerConnectionPipeline(
-        ServerIncomingConnection(incoming, outgoing, socket.remoteAddress),
+        ServerIncomingConnection(
+            incoming,
+            outgoing,
+            socket.remoteAddress
+        ),
         timeouts
     ) { request: Request ->
         val requestScope = this
