+++ ktor-server/ktor-server-benchmarks/src/jmh/kotlin/io/ktor/server/benchmarks/cio/CIOPlatformBenchmark.kt
@@ -8,6 +8,7 @@ import io.ktor.http.*
 import io.ktor.http.cio.*
 import io.ktor.server.benchmarks.*
 import io.ktor.server.cio.*
+import io.ktor.server.cio.backend.*
 import io.netty.util.*
 import kotlinx.coroutines.*
 import io.ktor.utils.io.*
