+++ ktor-server/ktor-server-cio/jvm/src/io/ktor/server/cio/CIOApplicationEngine.kt
@@ -5,6 +5,7 @@
 package io.ktor.server.cio
 
 import io.ktor.application.*
+import io.ktor.server.cio.backend.*
 import io.ktor.server.engine.*
 import io.ktor.util.*
 import io.ktor.util.pipeline.*
